pyproject.toml
[project]
name = "flexclaw3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agno[os]>=2.5.11",
    "anthropic>=0.86.0",
    "beautifulsoup4>=4.14.3",
    "chromadb>=0.4.0",
    "crawl4ai>=0.8.6",
    "ddgs>=9.12.0",
    "fastapi>=0.135.2",
    "lxml-html-clean>=0.4.4",
    "mcp>=1.26.0",
    "newspaper4k>=0.9.5",
    "onnxruntime<1.20",
    "openai>=2.30.0",
    "openpyxl>=3.1.0",
    "pypdf>=6.9.0",
    "python-docx>=1.1.0",
    "python-dotenv>=1.2.2",
    "python-pptx>=1.0.0",
    "python-telegram-bot>=21.0",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.48",
    "toolbox-core>=1.0.0",
    "uvicorn[standard]>=0.42.0",
    "wikipedia>=1.4.0",
    "youtube-transcript-api>=1.2.4",
    "playwright>=1.40.0",
    "fpdf2>=2.8.0",
    "markdown>=3.5.0",
    "cloudscraper>=1.2.71",
]
